threading issues fixes. Now types (concurrent hashmap) and activeTypeRequest (synchronized on) maps in the TypeInferences and InvariantTypeResources are thread safe.
Load type tries to synchronized  add the type if not it will wait until it can and then it first test the cached types so that it is not created twice..

problem was that activeTypeRequest was already filled with type "JSForm" and then it returns a null. That ended up in a unknonw type because of the proxy lookup that failed.