fix for stuff like this:
					fStartDate = fStartDate.addDays(1).clearTime();
					fEndDate = fStartDate.clone().addDays(1).addSeconds(-1);

fStartDate == new Date()
but Date is prototyped and have all kinds of functions added to it.
DLTK doesn't know them at this time (thats another feature request)
so if you have a few of these lines, it will constantly try to resolve them, because its the start entry point (fStartDate.addDays) and also the result.  (the result is assigned) and then later on that is again used, so you have deeply nested references pointing to each other.