different implementation of previous commit:
"this" should never be a reference but always be a copy so that changes
to the instance are not shown up on the this so also on the other
instances.


now a realy copy is only created when a new Xxx() is done. Then the
ThisValue of the function is copied instead of referenced, any thing
else it is still referenced.